fhirRouter tests
